requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
face_recognition>=1.3.0
PyYAML>=6.0
numpy>=1.24.0
mss>=9.0.0
Pillow>=10.0.0
